C: Promote function parameters from u8/s8/u16 to int throughout
Update all scalar-valued function parameters in ChaseHQ.c (and the
callback typedefs that constrain their signatures in ChaseHQ-State.h
and ChaseHQ-Internal.h) from u8/s8/u16 to int.  Pointer parameters
(const u8 *, u8 *, const u16 *) are unchanged.

Typedefs updated: dr_callback_t, plot_sprite_cb_t, dr_fill_fn,
obj_handler_t, dso_callback_t, draw_object_entrypt_t.

+++ C/ChaseHQ/ChaseHQ-Internal.h
@@ -97,7 +97,7 @@
 /* ----------------------------------------------------------------------- */
 
 typedef void obj_handler_t(chqstate_t *state,
-                           u8          Biterations,
+                           int         Biterations,
                            const void *arg,
                            const u16  *IXxpos,
                            const u8   *IYheight);
